requirements.txt
# Core dependencies for the document intelligence project
anthropic>=0.40.0
python-dotenv>=1.0.0

# You'll add these later as you expand:
# pytest>=8.0.0           # For proper testing
# pytest-asyncio>=0.23.0  # For async tests
# pydantic>=2.0.0         # For structured outputs
# pandas>=2.0.0           # For data analysis
# pymupdf>=1.24.0         # For PDF processing (fitz)
# mcp>=0.1.0              # For MCP server
